Implement ProjectClient for Project API calls
This collection of commits will focus on defining the `ProjectClient`
which is accessed through the `TrackBearClient` from the `project`
attribute.

During development, the `exceptions.py` and `models.py` modules have
been created to house shared objects for the library. These are
considered public API for imports, allowing the client to propery
annotate types as needed.