Restructure external API of TrackBearClient
This change restructures the internal use of `APIClient` to simplify the
namespaces of everything. This is a breaking change.

`TrackBearClient` and `ProjectClient` no longer inherit from
`APIClient`. Instead the `APIClient` is built on initialization of a
`TrackBearClient` instance and passed as a arguement to additional
clients.

The `.get`, `.post`, `.delete`, and `.patch` methods of
`TrackBearClient` are now accessed from `TrackBearClient.bare`.

The following methods have been renamed for simplicity in the
`TrackBearClient.project` provider:

* `.get_by_id()` -> `.get`
* `.delete_by_id()` -> `.delete`