Replace Literal use with enums
Replacing uses of `typing.Literal` with `enum.Enum`. This change adds
the `enums.py` module which is also imported and exposed at the public
api level. `Phase` and `State` string enums are added.

Using enums will reduce the overhead of the next-dev by providing a
clear definition of valid values. There will also be faster failures if
incoming values are invalid.

For the `ProjectClient` API, both `str` and `Phase` are accepted as
parameters. The incoming value is forced into a `Phase` enum before any
API calls for run-time validation.