Implement Goal Client
Ha! This one was fun. The Goal is a little different with some
complexity in the model depending on the type of goal. There are Target
goals, Habit goals, and Habit goals with targets.

Not 100% enjoying the `save_target()` and `save_habit()` methods but it
felt better than trying to fit both into one confusing signature.